Implemented all methods in Turtle in sync with Parser, cast all angles and magnitudes to int

+++ src/model/IState.java
@@ -15,7 +15,7 @@ public interface IState {
      * @return the absolute heading of the object on the perspective of 
      * the user
      */
-    public double getHeading();
+    public int getHeading();
     
     public boolean isPenWriting();
